+++ src/main/java/io/ix0rai/bodaciousberries/Bodaciousberries.java
@@ -1,12 +1,12 @@
 package io.ix0rai.bodaciousberries;
 
 import io.ix0rai.bodaciousberries.compat.Compat;
+import io.ix0rai.bodaciousberries.particle.Particles;
 import io.ix0rai.bodaciousberries.registry.Berries;
 import io.ix0rai.bodaciousberries.registry.BodaciousBlocks;
 import io.ix0rai.bodaciousberries.registry.Bushes;
 import io.ix0rai.bodaciousberries.registry.Juices;
 import io.ix0rai.bodaciousberries.registry.Sounds;
-import io.ix0rai.bodaciousberries.registry.particles.Particles;
 import io.ix0rai.bodaciousberries.worldgen.BerryBushPatchGen;
 import net.fabricmc.api.ModInitializer;
 import net.minecraft.util.Identifier;
