+++ src/main/java/io/ix0rai/bodaciousberries/block/RainberryBush.java
@@ -1,6 +1,6 @@
 package io.ix0rai.bodaciousberries.block;
 
-import io.ix0rai.bodaciousberries.registry.particles.Particles;
+import io.ix0rai.bodaciousberries.particle.Particles;
 import net.minecraft.block.BlockState;
 import net.minecraft.item.Item;
 import net.minecraft.util.math.BlockPos;
