+++ src/main/java/io/ix0rai/bodaciousberries/block/entity/BerryHarvesterBlockEntity.java
@@ -3,8 +3,8 @@ package io.ix0rai.bodaciousberries.block.entity;
 import io.ix0rai.bodaciousberries.block.BasicBerryBush;
 import io.ix0rai.bodaciousberries.block.BerryBush;
 import io.ix0rai.bodaciousberries.block.BerryHarvesterBlock;
+import io.ix0rai.bodaciousberries.particle.Particles;
 import io.ix0rai.bodaciousberries.registry.BodaciousBlocks;
-import io.ix0rai.bodaciousberries.registry.particles.Particles;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.SweetBerryBushBlock;
