+++ src/main/java/io/ix0rai/bodaciousberries/registry/Berries.java
@@ -4,9 +4,9 @@ import io.ix0rai.bodaciousberries.Bodaciousberries;
 import io.ix0rai.bodaciousberries.block.BerryBush;
 import io.ix0rai.bodaciousberries.block.DoubleBerryBush;
 import io.ix0rai.bodaciousberries.block.GrowingBerryBush;
-import io.ix0rai.bodaciousberries.registry.items.ChorusBerries;
-import io.ix0rai.bodaciousberries.registry.items.GojiBerries;
-import io.ix0rai.bodaciousberries.registry.items.Rainberry;
+import io.ix0rai.bodaciousberries.item.ChorusBerries;
+import io.ix0rai.bodaciousberries.item.GojiBerries;
+import io.ix0rai.bodaciousberries.item.Rainberry;
 import net.minecraft.item.AliasedBlockItem;
 import net.minecraft.item.FoodComponent;
 import net.minecraft.item.Item;
